Refactored code to be more clean
Took a cue from Lou and the original ess_api.rb "smelled" too bad.
Refactored into different classes for portability and ease of use.

requests_controller.rb - All that was done here was to change from using
the module to an instantiated object.

appliance_web_request.rb - This class performs the formation of a web
request, with OAuth signing and equivalent.

appliance_xml_parser - Parses XML received into JSON response. This
class will need to change if XML specs from the appliance or upper
management needs to know more about the system

capture_appliance.rb - Model for representation of a capture appliance.
capture_status_monitor.rb Capture status monitor object. Kind of a
wrapper for capture_appliance. Will do more stuff in the future.

bypass_ssl_verification_for_oauth.rb - because we are cheapskates who
don't want to splurge for a unique SSL cert for each of the 40-something
recording appliances we have on campus, this is necessary to override
default SSL behavior of VERIFY_PEER to VERIFY_NONE. It makes MIM attacks
easier, but we aren't transmitting any sensitive data so it should be
ok.

app/helpers/ess_api.rb should be considered deprecated and is left in
only for reference.